+++ graylog2-web-interface/src/views/components/searchbar/time-range-filter/time-range-picker/RelativeRangeSelect.tsx
@@ -21,7 +21,7 @@ import moment from 'moment';
 
 import { RELATIVE_RANGE_TYPES } from 'views/Constants';
 import { Select } from 'components/common';
-import type { RangeClassified } from 'views/components/searchbar/time-range-picker/types';
+import type { RangeClassified } from 'views/components/searchbar/time-range-filter/time-range-picker/types';
 
 import { isTypeRelativeClassified, RELATIVE_CLASSIFIED_ALL_TIME_RANGE } from './RelativeTimeRangeClassifiedHelper';
 import RelativeRangeValueInput from './RelativeRangeValueInput';
