+++ graylog2-web-interface/src/views/components/searchbar/time-range-filter/time-range-picker/TabAbsoluteTimeRange.tsx
@@ -26,7 +26,7 @@ import type { AbsoluteTimeRange } from 'views/logic/queries/Query';
 import useUserDateTime from 'hooks/useUserDateTime';
 import TimeRangeInputSettingsContext from 'views/components/contexts/TimeRangeInputSettingsContext';
 import useSearchConfiguration from 'hooks/useSearchConfiguration';
-import TabPresetDropdown from 'views/components/searchbar/time-range-picker/TabPresetDropdown';
+import TabPresetDropdown from 'views/components/searchbar/time-range-filter/time-range-picker/TabPresetDropdown';
 
 import AbsoluteCalendar from './AbsoluteCalendar';
 import AbsoluteTimestamp from './AbsoluteTimestamp';
