+++ graylog2-web-interface/src/views/components/searchbar/time-range-filter/time-range-picker/TimeRangeAddToQuickListButton.tsx
