+++ graylog2-web-interface/src/views/components/searchbar/time-range-filter/time-range-picker/TimeRangeLivePreview.tsx
@@ -25,11 +25,11 @@ import { isTypeRelative, isTypeRelativeWithEnd, isTypeRelativeWithStartOnly } fr
 import type { TimeRange, NoTimeRangeOverride } from 'views/logic/queries/Query';
 import { Icon, IfPermitted } from 'components/common';
 import { DATE_TIME_FORMATS } from 'util/DateTime';
-import TimeRangeAddToQuickListButton from 'views/components/searchbar/time-range-picker/TimeRangeAddToQuickListButton';
-import type { TimeRangeDropDownFormValues } from 'views/components/searchbar/time-range-picker/TimeRangePicker';
+import TimeRangeAddToQuickListButton from 'views/components/searchbar/time-range-filter/time-range-picker/TimeRangeAddToQuickListButton';
+import type { TimeRangeDropDownFormValues } from 'views/components/searchbar/time-range-filter/time-range-picker/TimeRangePicker';
 import TimeRangeInputSettingsContext from 'views/components/contexts/TimeRangeInputSettingsContext';
 
-import { EMPTY_OUTPUT, EMPTY_RANGE } from '../time-range-filter/TimeRangeDisplay';
+import { EMPTY_OUTPUT, EMPTY_RANGE } from '../TimeRangeDisplay';
 
 type Props = {
   timerange?: TimeRange | NoTimeRangeOverride,
